+++ packages/graphql/src/schema/resolvers/wrapper.ts
@@ -29,12 +29,11 @@ import type { ExecutorConstructorParam } from "../../classes/Executor";
 import { DEBUG_GRAPHQL } from "../../constants";
 import createAuthParam from "../../translate/create-auth-param";
 import type { Context, Neo4jAuthorizationSettings, Neo4jGraphQLPlugins } from "../../types";
-import { getToken, parseBearerToken } from "../../utils/get-token";
 import type { SubscriptionConnectionContext, SubscriptionContext } from "./subscriptions/types";
 import { decodeToken, verifyGlobalAuthentication } from "./wrapper-utils";
 import type { Neo4jGraphQLSchemaModel } from "../../schema-model/Neo4jGraphQLSchemaModel";
 import { IncomingMessage } from "http";
-import { Neo4jGraphQLAuthorization } from "../../classes/Neo4jGraphQLAuthorization";
+import { Neo4jGraphQLAuthorization, getToken, parseBearerToken } from "../../classes/Neo4jGraphQLAuthorization";
 
 const debug = Debug(DEBUG_GRAPHQL);
 
@@ -92,7 +91,7 @@ export const wrapResolver =
 
         if (!context.jwt) {
             if (authorization) {
-                const jwt = await new Neo4jGraphQLAuthorization(authorization).parse(context);
+                const jwt = await new Neo4jGraphQLAuthorization(authorization).parseFrom({ context });
                 jwt && (context.jwt = jwt);
             } else {
                 // TODO: remove this else after migrating to new authorization constructor
@@ -161,13 +160,13 @@ export const wrapSubscription =
         };
 
         if (!context?.jwt && contextParams.authorization) {
-            const token = parseBearerToken(contextParams.authorization);
             if (resolverArgs.authorization) {
-                subscriptionContext.jwt = new Neo4jGraphQLAuthorization(resolverArgs.authorization).decodeWithoutVerify(
-                    token
-                );
+                subscriptionContext.jwt = new Neo4jGraphQLAuthorization(resolverArgs.authorization).parseFrom({
+                    bearerToken: contextParams.authorization,
+                });
             } else {
                 // TODO: remove this else after migrating to new authorization constructor
+                const token = parseBearerToken(contextParams.authorization);
                 subscriptionContext.jwt = await decodeToken(token, plugins.auth);
             }
         }
